refactor fixture setup
