+++ testCafe/src/common/fixtureSetup.ts
@@ -2,9 +2,9 @@ import dotenv from 'dotenv';
 
 dotenv.config();
 
-const setupFixture: any = (ENV_URL: string) => {
+const setupFixture: any = (description: string, ENV_URL: string) => {
   const WEBSITE_URL = process.env[ENV_URL];
-  return fixture`Home page test`.page`${WEBSITE_URL}`;
+  return fixture(description).page(`${WEBSITE_URL}`);
 };
 
 export default setupFixture;
